+++ SRC/Injector/Private/ServiceCollection.cs
@@ -4,37 +4,51 @@
 * Author: Denes Solti                                                           *
 ********************************************************************************/
 using System;
+using System.Collections;
 using System.Collections.Generic;
 
 namespace Solti.Utils.DI.Internals
 {
     using Interfaces;
+    using Properties;
 
-    internal sealed class ServiceCollection : HashSet<AbstractServiceEntry>, IModifiedServiceCollection
+    internal sealed class ServiceCollection : IModifiedServiceCollection
     {
+        private readonly HashSet<AbstractServiceEntry> FUnderlyingCollection = new(ServiceIdComparer.Instance);
         private AbstractServiceEntry? FLastEntry;
 
-        //
-        // TODO: FIXME: new ServiceCollection().Add() won't call this method...
-        //
-
-        bool ISet<AbstractServiceEntry>.Add(AbstractServiceEntry item)
+        public void Add(AbstractServiceEntry item)
         {
             if (item is null)
                 throw new ArgumentNullException(nameof(item));
 
-            bool success = Add(item);
-            if (success)
-                FLastEntry = item;
-            
-            return success;
+            if (!FUnderlyingCollection.Add(item ?? throw new ArgumentNullException(nameof(item))))
+                throw new ServiceAlreadyRegisteredException(Resources.SERVICE_ALREADY_REGISTERED);
+  
+            FLastEntry = item;
         }
 
-        public ServiceCollection(ServiceOptions? serviceOptions = null) : base(ServiceIdComparer.Instance)
+        public void Clear() => FUnderlyingCollection.Clear();
+
+        public bool Contains(AbstractServiceEntry item) => FUnderlyingCollection.Contains(item ?? throw new ArgumentNullException(nameof(item)));
+
+        public void CopyTo(AbstractServiceEntry[] array, int arrayIndex) => FUnderlyingCollection.CopyTo(array ?? throw new ArgumentNullException(nameof(array)), arrayIndex);
+
+        public bool Remove(AbstractServiceEntry item) => FUnderlyingCollection.Remove(item ?? throw new ArgumentNullException(nameof(item)));
+
+        public IEnumerator<AbstractServiceEntry> GetEnumerator() => FUnderlyingCollection.GetEnumerator();
+
+        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+
+        public ServiceCollection(ServiceOptions? serviceOptions = null) : base()
             => ServiceOptions = serviceOptions ?? ServiceOptions.Default;
 
         public AbstractServiceEntry LastEntry => FLastEntry ?? throw new InvalidOperationException();
 
         public ServiceOptions ServiceOptions { get; }
+
+        public int Count => FUnderlyingCollection.Count;
+
+        public bool IsReadOnly { get; }
     }
 }