+++ SRC/Injector/Properties/Resources.Designer.cs
@@ -232,6 +232,15 @@ namespace Solti.Utils.DI.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to A service with the same ID has already been registered..
+        /// </summary>
+        internal static string SERVICE_ALREADY_REGISTERED {
+            get {
+                return ResourceManager.GetString("SERVICE_ALREADY_REGISTERED", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Service &quot;{0}&quot; could not be found..
         /// </summary>
         internal static string SERVICE_NOT_FOUND {